better

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: ialvarez <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/08 20:11:04 by ialvarez          #+#    #+#             */
-/*   Updated: 2022/02/02 06:09:21 by ialvarez         ###   ########.fr       */
+/*   Updated: 2022/02/08 21:33:23 by ialvarez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int		i;
 	long	res;
